requirements.txt
Django==5.1.2
python-telegram-bot==21.6
python-dotenv==1.0.1
